Retreive CaseComment CustomObject and Workflow
Salesforce does not return a `FileProperties` instance for
`CaseComment` when 'listMetadata` query is sent for `CustomObject` or
`Workflow` types, resulting in a backup that is missing these metadata
components.

This fix works around that by manually adding stub `FileProperties`
instances for `CaseComment` object and workflow, so that the metadata can
be retrieved.

Fixes #46