Filter out Apex Classes from managed packages
When retrieving ApexClass metadata type, do not retrieve components from
manage packages. When Apex classes from managed packages are retrieved
the file contents is hidden, so there is nothing to backup.

Also, in large Orgs, there can be more that 10000 components which can
result in LIMIT_EXCEPTION upon retrieval.

Now the list of Apex classes to be retrieved will be filtered, so only
classes with no namespace prefix will be retrieved.