Wildcard retrieve triggers, components and lwc
Change the following components to be retrieved with wildcard in
package.xml manifest instead of using <sf:bulkRetrieve>:
 - Apex Triggers
 - Visualforce Compoents
 - Lightning Web Components,

When the above metadata types are retrieved with <sf:bulkRetrieve>,
the retrieval will include components that belong to managed packages
and that have their content hidden. Since these files contains have no
content for backup, it's a waste of time retrieving them. So instead we
will use the wildcard in package.xml for these metadata types, which
will limit components to non managed package components.