Support Salesforce Session ID
Support running the backup with a `sf.sessionId` property, so that an
existing Salesforce Session ID or OAuth token can be used for login.

Reuse the same Salesforce session for all retreive tasks, to avoid
unnecessarily hitting the login server with each retrieve task and avoid
running into a LOGIN_RATE_EXCEEDED error.

When a username/password are used for authentication, we will now
initialize a Salesforce session and use that session id for the duration
of the metadata backup too.

Also:
 - Bundles groovy dependencies for ant to use `<groovy>` and `<groovyc>`
   ant tasks.
 - `force-meta-backup.groovy` script is no longer an excutable CLI
   script, remove commons-cli dependency. Script now just provides
   helpers for ant to execute via `<groovy>` task.
 - Groovy source code `force-meta-backup.groovy` is compiled using
   `<groovyc>`.
 - Groovy helpers are invoked using `<groovy>` task instead `<exec>`.
 - Fixed warning where `ArrayList.size` field was being accessed by
   Groovy, calling `ArrayList.size()` now instead.

Fixes #76